Move audit logs types
