+++ workos/resources/audit_logs.py
@@ -2,7 +2,6 @@ from typing import Any, Literal, Mapping, Optional
 from workos.resources.workos_model import WorkOSModel
 
 AuditLogExportState = Literal["error", "pending", "ready"]
-AuditLogMetadata = Mapping[str, Any]
 
 
 class AuditLogExport(WorkOSModel):
