+++ workos/types/audit_logs/audit_log_event_context.py
@@ -1,0 +1,8 @@
+from typing import NotRequired, TypedDict
+
+
+class AuditLogEventContext(TypedDict):
+    """Attributes of audit log event context."""
+
+    location: str
+    user_agent: NotRequired[str]
