+++ workos/types/audit_logs/audit_log_metadata.py
@@ -1,0 +1,4 @@
+from typing import Any, Mapping
+
+
+AuditLogMetadata = Mapping[str, Any]
